+++ src/actions/navigationActions.js
@@ -4,6 +4,10 @@ export function resetAllAction() {
   return {type: 'FULL_RESET'};
 }
 
+export function goToHomePage() {
+  return push('/');
+}
+
 export function goToMusicGenre(slug) {
   return push(`/${slug}`);
 }
@@ -12,10 +16,15 @@ export function search(query) {
   return push(`/search/${query}`);
 }
 
+export function goToRandomPage() {
+  return push('/random-musical-genre');
+}
+
 export function goToAddVideo() {
   return push(`/add/video`);
 }
 
 export function goToAddVideoToGenre(genreSlug, lockGenre) {
-  return push(`/add/video/${genreSlug}` + (lockGenre ? '/lock' : ''));
+  const encodedGenre = encodeURIComponent(genreSlug);
+  return push(`/add/video/${encodedGenre}` + (lockGenre ? '/lock' : ''));
 }