+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+
+- Add a page that lists all musical genres
+- Random genre is now a page on its own
+
+#v1.0.1
+
+- Fix genre names with slashes
+
 #v1.0.0
 
 - First Major Release